Update packages to reflect api/impl module split
Move files to packages matching module structure:
- API classes to cz.adamec.timotej.snag.lib.sync.fe.app.api
- Handler classes to cz.adamec.timotej.snag.lib.sync.fe.app.api.handler
- Impl classes to cz.adamec.timotej.snag.lib.sync.fe.app.impl.internal
- DI module to cz.adamec.timotej.snag.lib.sync.fe.app.impl.di

Also restore KDoc to EnqueueSyncSaveUseCase and EnqueueSyncDeleteUseCase.
