+++ feat/findings/fe/driven/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/findings/fe/driven/internal/sync/FindingSyncHandler.kt
@@ -18,7 +18,7 @@ import cz.adamec.timotej.snag.findings.fe.ports.FindingsApi
 import cz.adamec.timotej.snag.findings.fe.ports.FindingsDb
 import cz.adamec.timotej.snag.lib.core.fe.OfflineFirstDataResult
 import cz.adamec.timotej.snag.lib.core.fe.OnlineDataResult
-import cz.adamec.timotej.snag.lib.sync.fe.app.handler.DbApiSyncHandler
+import cz.adamec.timotej.snag.lib.sync.fe.app.api.handler.DbApiSyncHandler
 import kotlinx.coroutines.flow.Flow
 import kotlin.uuid.Uuid
