+++ feat/projects/fe/driven/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/projects/fe/driven/internal/sync/ProjectSyncHandler.kt
@@ -14,7 +14,7 @@ package cz.adamec.timotej.snag.projects.fe.driven.internal.sync
 
 import cz.adamec.timotej.snag.lib.core.fe.OfflineFirstDataResult
 import cz.adamec.timotej.snag.lib.core.fe.OnlineDataResult
-import cz.adamec.timotej.snag.lib.sync.fe.app.handler.DbApiSyncHandler
+import cz.adamec.timotej.snag.lib.sync.fe.app.api.handler.DbApiSyncHandler
 import cz.adamec.timotej.snag.projects.business.Project
 import cz.adamec.timotej.snag.projects.fe.driven.internal.LH
 import cz.adamec.timotej.snag.projects.fe.ports.ProjectsApi
