+++ feat/structures/fe/driven/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/structures/fe/driven/di/StructuresDrivenModule.kt
@@ -13,7 +13,7 @@
 package cz.adamec.timotej.snag.structures.fe.driven.di
 
 import cz.adamec.timotej.snag.lib.core.common.di.getIoDispatcher
-import cz.adamec.timotej.snag.lib.sync.fe.app.handler.SyncOperationHandler
+import cz.adamec.timotej.snag.lib.sync.fe.app.api.handler.SyncOperationHandler
 import cz.adamec.timotej.snag.structures.fe.driven.internal.api.RealStructuresApi
 import cz.adamec.timotej.snag.structures.fe.driven.internal.db.RealStructuresDb
 import cz.adamec.timotej.snag.structures.fe.driven.internal.sync.RealStructuresSync
