+++ lib/sync/fe/app/api/src/commonMain/kotlin/cz/adamec/timotej/snag/lib/sync/fe/app/api/EnqueueSyncDeleteUseCase.kt
@@ -10,11 +10,21 @@
  * Department of Software Engineering
  */
 
-package cz.adamec.timotej.snag.lib.sync.fe.app
+package cz.adamec.timotej.snag.lib.sync.fe.app.api
 
 import kotlin.uuid.Uuid
 
+/**
+ * Enqueues a delete sync operation for the given entity.
+ *
+ * Make sure a [cz.adamec.timotej.snag.lib.sync.fe.app.api.handler.SyncOperationHandler] is registered
+ * for the given [entityTypeId].
+ */
 interface EnqueueSyncDeleteUseCase {
+    /**
+     * @throws IllegalArgumentException if [cz.adamec.timotej.snag.lib.sync.fe.app.api.handler.SyncOperationHandler]
+     * is not registered for given [entityTypeId].
+     */
     suspend operator fun invoke(
         entityTypeId: String,
         entityId: Uuid,
