+++ lib/sync/fe/app/api/src/commonMain/kotlin/cz/adamec/timotej/snag/lib/sync/fe/app/api/handler/SyncOperationHandler.kt
@@ -10,15 +10,16 @@
  * Department of Software Engineering
  */
 
-package cz.adamec.timotej.snag.lib.sync.fe.app.internal
+package cz.adamec.timotej.snag.lib.sync.fe.app.api.handler
 
 import cz.adamec.timotej.snag.lib.sync.business.SyncOperationType
 import kotlin.uuid.Uuid
 
-internal interface EnqueueSyncOperationUseCase {
-    suspend operator fun invoke(
-        entityTypeId: String,
+interface SyncOperationHandler {
+    val entityTypeId: String
+
+    suspend fun execute(
         entityId: Uuid,
         operationType: SyncOperationType,
-    )
+    ): SyncOperationResult
 }
