--- lib/sync/fe/app/api/src/commonMain/kotlin/cz/adamec/timotej/snag/lib/sync/fe/app/handler/SyncOperationResult.kt
@@ -1,21 +1,0 @@
-/*
- * Copyright (c) 2026 Timotej Adamec
- * SPDX-License-Identifier: MIT
- *
- * This file is part of the thesis:
- * "Multiplatform snagging system with code sharing maximisation"
- *
- * Czech Technical University in Prague
- * Faculty of Information Technology
- * Department of Software Engineering
- */
-
-package cz.adamec.timotej.snag.lib.sync.fe.app.handler
-
-sealed interface SyncOperationResult {
-    data object Success : SyncOperationResult
-
-    data object EntityNotFound : SyncOperationResult
-
-    data object Failure : SyncOperationResult
-}
