+++ lib/sync/fe/app/impl/src/commonTest/kotlin/cz/adamec/timotej/snag/lib/sync/fe/app/impl/SyncEngineTest.kt
@@ -10,13 +10,13 @@
  * Department of Software Engineering
  */
 
-package cz.adamec.timotej.snag.lib.sync.fe.app
+package cz.adamec.timotej.snag.lib.sync.fe.app.impl
 
 import cz.adamec.timotej.snag.lib.core.common.ApplicationScope
 import cz.adamec.timotej.snag.lib.sync.business.SyncOperationType
-import cz.adamec.timotej.snag.lib.sync.fe.app.handler.SyncOperationHandler
-import cz.adamec.timotej.snag.lib.sync.fe.app.handler.SyncOperationResult
-import cz.adamec.timotej.snag.lib.sync.fe.app.internal.SyncEngine
+import cz.adamec.timotej.snag.lib.sync.fe.app.api.handler.SyncOperationHandler
+import cz.adamec.timotej.snag.lib.sync.fe.app.api.handler.SyncOperationResult
+import cz.adamec.timotej.snag.lib.sync.fe.app.impl.internal.SyncEngine
 import cz.adamec.timotej.snag.lib.sync.fe.driven.test.FakeSyncQueue
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
